+++ pandas/core/tools/timedeltas.py
