+++ pandas/tests/series/test_combine_concat.py
@@ -204,7 +204,7 @@ class TestSeriesCombine(TestData, tm.TestCase):
         self.assertEqual(result.ftype, 'object:dense')
 
     def test_combine_first_dt64(self):
-        from pandas.tseries.tools import to_datetime
+        from pandas.core.tools.datetimes import to_datetime
         s0 = to_datetime(Series(["2010", np.NaN]))
         s1 = to_datetime(Series([np.NaN, "2011"]))
         rs = s0.combine_first(s1)
