+++ pandas/tests/tseries/test_frequencies.py
@@ -7,7 +7,7 @@ from pandas import (Index, DatetimeIndex, Timestamp, Series,
                     date_range, period_range)
 
 import pandas.tseries.frequencies as frequencies
-from pandas.tseries.tools import to_datetime
+from pandas.core.tools.datetimes import to_datetime
 
 import pandas.tseries.offsets as offsets
 from pandas.core.indexes.period import PeriodIndex
