+++ pandas/tseries/api.py
@@ -5,5 +5,4 @@
 # flake8: noqa
 
 from pandas.tseries.frequencies import infer_freq
-from pandas.tseries.timedeltas import to_timedelta
 import pandas.tseries.offsets as offsets
